+++ docs/source/overview/core-concepts/physical-backends/newton/newton-manager-abstraction.rst
@@ -176,9 +176,9 @@ You can exercise this coupling path with the Franka soft-body lifting task:
 
 .. code-block:: bash
 
-   ./isaaclab.sh -p scripts/environments/zero_agent.py --task Isaac-Lift-Soft-Franka-v0 --num_envs 1 --visualizer kit
+   ./isaaclab.sh -p scripts/environments/zero_agent.py --task Isaac-Lift-Soft-Franka --num_envs 1 --visualizer kit
 
-For the surface-deformable cloth variant, use ``--task Isaac-Lift-Cloth-Franka-v0``.
+For the surface-deformable cloth variant, use ``--task Isaac-Lift-Cloth-Franka``.
 
 
 This environment configures
