+++ scripts/environments/state_machine/lift_franka_soft.py
@@ -24,7 +24,7 @@ from isaaclab.app import AppLauncher
 # add argparse arguments
 parser = argparse.ArgumentParser(description="Pick and lift a deformable with a robotic arm.")
 parser.add_argument("--num_envs", type=int, default=1, help="Number of environments to simulate.")
-parser.add_argument("--task", type=str, default="Isaac-Lift-Soft-Franka-v0", help="The task to run.")
+parser.add_argument("--task", type=str, default="Isaac-Lift-Soft-Franka", help="The task to run.")
 parser.add_argument("--video", action="store_true", default=False, help="Record a video of the rollout.")
 parser.add_argument("--video_length", type=int, default=500, help="Length of the recorded video (in env steps).")
 parser.add_argument(
@@ -63,8 +63,8 @@ import warp as wp
 from isaaclab.assets.deformable_object.deformable_object_data import DeformableObjectData
 
 import isaaclab_tasks  # noqa: F401
-from isaaclab_tasks.manager_based.manipulation.lift_franka_soft.franka_cloth_env_cfg import FrankaClothEnvCfg
-from isaaclab_tasks.manager_based.manipulation.lift_franka_soft.franka_soft_env_cfg import FrankaSoftEnvCfg
+from isaaclab_tasks.manager_based.manipulation.lift.config.franka_soft.franka_cloth_env_cfg import FrankaClothEnvCfg
+from isaaclab_tasks.manager_based.manipulation.lift.config.franka_soft.franka_soft_env_cfg import FrankaSoftEnvCfg
 from isaaclab_tasks.utils.parse_cfg import parse_env_cfg
 
 # initialize warp
@@ -284,24 +284,24 @@ class PickAndLiftSm:
 def main():
     # create environment
     render_mode = "rgb_array" if args_cli.video else None
-    if args_cli.task == "Isaac-Lift-Soft-Franka-v0":
+    if args_cli.task == "Isaac-Lift-Soft-Franka":
         # parse configuration
         env_cfg: FrankaSoftEnvCfg = parse_env_cfg(
-            "Isaac-Lift-Soft-Franka-v0",
+            "Isaac-Lift-Soft-Franka",
             device=args_cli.device,
             num_envs=args_cli.num_envs,
         )
         env_cfg.viewer.eye = (2.1, 1.0, 1.3)
-        env = gym.make("Isaac-Lift-Soft-Franka-v0", cfg=env_cfg, render_mode=render_mode)
-    elif args_cli.task == "Isaac-Lift-Soft-Franka-Cloth-v0":
+        env = gym.make("Isaac-Lift-Soft-Franka", cfg=env_cfg, render_mode=render_mode)
+    elif args_cli.task == "Isaac-Lift-Cloth-Franka":
         # parse configuration
         env_cfg: FrankaClothEnvCfg = parse_env_cfg(
-            "Isaac-Lift-Cloth-Franka-v0",
+            "Isaac-Lift-Cloth-Franka",
             device=args_cli.device,
             num_envs=args_cli.num_envs,
         )
         env_cfg.viewer.eye = (2.1, 1.0, 1.3)
-        env = gym.make("Isaac-Lift-Cloth-Franka-v0", cfg=FrankaClothEnvCfg(), render_mode=render_mode)
+        env = gym.make("Isaac-Lift-Cloth-Franka", cfg=FrankaClothEnvCfg(), render_mode=render_mode)
     else:
         raise ValueError(f"Unknown task: {args_cli.task}")
