+++ source/isaaclab/test/envs/test_action_state_recorder_term.py
@@ -83,7 +83,7 @@ def check_initial_state_recorder_term(env):
         assert are_states_equal, output_log
 
 
-@pytest.mark.parametrize("task_name", ["Isaac-Lift-Cube-Franka-v0"])
+@pytest.mark.parametrize("task_name", ["Isaac-Lift-Cube-Franka"])
 @pytest.mark.parametrize("device", ["cuda:0", "cpu"])
 @pytest.mark.parametrize("num_envs", [1, 2])
 def test_action_state_recorder_terms(task_name, device, num_envs, temp_dir):
