+++ source/isaaclab_rl/changelog.d/mhaiderbhai-merge-lift-tasks.skip
@@ -1,0 +1,1 @@
+Test-only: updated lift task IDs in tests to match the consolidated/renamed environments.
