+++ source/isaaclab_rl/test/export/test_rsl_rl_export_flow.py
@@ -83,8 +83,8 @@ TASKS = [
     "Isaac-Reach-UR10-v0",
     "Isaac-Reach-UR10-Play-v0",
     # Manipulation Lift
-    "Isaac-Lift-Cube-Franka-v0",
-    "Isaac-Lift-Cube-Franka-Play-v0",
+    "Isaac-Lift-Cube-Franka",
+    "Isaac-Lift-Cube-Franka-Play",
     # Manipulation Cabinet
     "Isaac-Open-Drawer-Franka-v0",
     "Isaac-Open-Drawer-Franka-Play-v0",
