+++ source/isaaclab_tasks/isaaclab_tasks/manager_based/manipulation/lift/config/franka_soft/__init__.py
@@ -11,9 +11,8 @@ from . import agents
 # Register Gym environments.
 ##
 
-
 gym.register(
-    id="Isaac-Lift-Soft-Franka-v0",
+    id="Isaac-Lift-Soft-Franka",
     entry_point="isaaclab.envs:ManagerBasedRLEnv",
     disable_env_checker=True,
     kwargs={
@@ -23,7 +22,7 @@ gym.register(
 )
 
 gym.register(
-    id="Isaac-Lift-Cloth-Franka-v0",
+    id="Isaac-Lift-Cloth-Franka",
     entry_point="isaaclab.envs:ManagerBasedRLEnv",
     disable_env_checker=True,
     kwargs={
