+++ source/isaaclab_tasks/isaaclab_tasks/manager_based/manipulation/lift/mdp/__init__.pyi
@@ -4,14 +4,43 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 __all__ = [
+    # observations
     "object_position_in_robot_root_frame",
+    "deformable_com_in_robot_root_frame",
+    "DeformableSampledPointsInRobotRootFrame",
+    # rewards
     "object_ee_distance",
     "object_goal_distance",
     "object_is_lifted",
+    "deformable_lifted",
+    "deformable_ee_distance",
+    "deformable_com_goal_distance",
+    "gripper_close_action",
+    # terminations
     "object_reached_goal",
+    "deformable_com_below_minimum",
+    "deformable_outside_table_bounds",
+    "ee_below_minimum",
 ]
 
-from .observations import object_position_in_robot_root_frame
-from .rewards import object_ee_distance, object_goal_distance, object_is_lifted
-from .terminations import object_reached_goal
+from .observations import (
+    DeformableSampledPointsInRobotRootFrame,
+    deformable_com_in_robot_root_frame,
+    object_position_in_robot_root_frame,
+)
+from .rewards import (
+    deformable_com_goal_distance,
+    deformable_ee_distance,
+    deformable_lifted,
+    gripper_close_action,
+    object_ee_distance,
+    object_goal_distance,
+    object_is_lifted,
+)
+from .terminations import (
+    deformable_com_below_minimum,
+    deformable_outside_table_bounds,
+    ee_below_minimum,
+    object_reached_goal,
+)
 from isaaclab.envs.mdp import *
