+++ source/isaaclab_tasks/isaaclab_tasks/manager_based/manipulation/lift/mdp/rewards.py
@@ -3,17 +3,20 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
+"""Reward functions for the rigid and deformable lift tasks."""
+
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
 import torch
+import warp as wp
 
 from isaaclab.managers import ManagerTermBase, RewardTermCfg, SceneEntityCfg
 from isaaclab.utils.math import combine_frame_transforms
 
 if TYPE_CHECKING:
-    from isaaclab.assets import RigidObject
+    from isaaclab.assets import Articulation, DeformableObject, RigidObject
     from isaaclab.envs import ManagerBasedRLEnv
     from isaaclab.sensors import FrameTransformer
 
@@ -90,3 +93,91 @@ class object_goal_distance(ManagerTermBase):
         if success_threshold is not None:
             self._succeeded |= is_lifted & (distance < success_threshold)
         return is_lifted.float() * (1 - torch.tanh(distance / std))
+
+
+def deformable_lifted(
+    env: ManagerBasedRLEnv,
+    minimal_height: float,
+    asset_cfg: SceneEntityCfg = SceneEntityCfg("deformable"),
+) -> torch.Tensor:
+    """Reward if the deformable COM is above a minimum height.
+
+    Args:
+        env: The environment instance.
+        minimal_height: Minimum COM height [m].
+        asset_cfg: The deformable object entity.
+
+    Returns:
+        Reward tensor with shape ``(num_envs,)``.
+    """
+    asset: DeformableObject = env.scene[asset_cfg.name]
+    com_z = wp.to_torch(asset.data.root_pos_w)[:, 2]
+    return torch.where(com_z > minimal_height, 1.0, 0.0)
+
+
+def deformable_ee_distance(
+    env: ManagerBasedRLEnv,
+    std: float,
+    asset_cfg: SceneEntityCfg = SceneEntityCfg("deformable"),
+    ee_frame_cfg: SceneEntityCfg = SceneEntityCfg("ee_frame"),
+) -> torch.Tensor:
+    """Reward reaching the deformable's nearest nodal point with the end-effector.
+
+    Args:
+        env: The environment instance.
+        std: The tanh kernel standard deviation [m].
+        asset_cfg: The deformable object entity.
+        ee_frame_cfg: The end-effector frame entity.
+
+    Returns:
+        Reward tensor with shape ``(num_envs,)``.
+    """
+    asset: DeformableObject = env.scene[asset_cfg.name]
+    ee_frame: FrameTransformer = env.scene[ee_frame_cfg.name]
+    nodal_pos_w = wp.to_torch(asset.data.nodal_pos_w)
+    ee_w = wp.to_torch(ee_frame.data.target_pos_w)[..., 0, :]
+    distance = torch.linalg.norm(nodal_pos_w - ee_w.unsqueeze(1), dim=2).min(dim=1).values
+    return 1.0 - torch.tanh(distance / std)
+
+
+def deformable_com_goal_distance(
+    env: ManagerBasedRLEnv,
+    std: float,
+    minimal_height: float,
+    command_name: str,
+    robot_cfg: SceneEntityCfg = SceneEntityCfg("robot"),
+    asset_cfg: SceneEntityCfg = SceneEntityCfg("deformable"),
+) -> torch.Tensor:
+    """Reward tracking of the goal position by the deformable's COM (tanh kernel).
+
+    Only credits when the COM is above ``minimal_height`` (i.e. the object is lifted).
+    The command is interpreted as ``[x, y, z, qw, qx, qy, qz]`` in the robot's root frame.
+    """
+    robot: Articulation = env.scene[robot_cfg.name]
+    asset: DeformableObject = env.scene[asset_cfg.name]
+    command = env.command_manager.get_command(command_name)
+    des_pos_b = command[:, :3]
+    des_pos_w, _ = combine_frame_transforms(
+        wp.to_torch(robot.data.root_pos_w), wp.to_torch(robot.data.root_quat_w), des_pos_b
+    )
+    com_w = wp.to_torch(asset.data.root_pos_w)
+    distance = torch.linalg.norm(des_pos_w - com_w, dim=1)
+    return (com_w[:, 2] > minimal_height) * (1.0 - torch.tanh(distance / std))
+
+
+def gripper_close_action(env: ManagerBasedRLEnv, action_name: str = "gripper_action") -> torch.Tensor:
+    """Penalty signal for commanding the gripper to close.
+
+    The binary gripper action uses negative float actions for close commands and
+    non-negative actions for open commands.
+
+    Args:
+        env: The environment instance.
+        action_name: Name of the gripper action term.
+
+    Returns:
+        Tensor with shape ``(num_envs,)`` containing ``1`` when the gripper is
+        commanded closed and ``0`` otherwise.
+    """
+    gripper_action = env.action_manager.get_term(action_name).raw_actions
+    return torch.any(gripper_action < 0.0, dim=1).float()
