+++ source/isaaclab_tasks/isaaclab_tasks/manager_based/manipulation/lift/mdp/terminations.py
@@ -3,7 +3,7 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
-"""Common functions that can be used to activate certain terminations for the lift task.
+"""Termination functions for the rigid and deformable lift tasks.
 
 The functions can be passed to the :class:`isaaclab.managers.TerminationTermCfg` object to enable
 the termination introduced by the function.
@@ -14,13 +14,15 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 import torch
+import warp as wp
 
 from isaaclab.managers import SceneEntityCfg
 from isaaclab.utils.math import combine_frame_transforms
 
 if TYPE_CHECKING:
-    from isaaclab.assets import RigidObject
+    from isaaclab.assets import DeformableObject, RigidObject
     from isaaclab.envs import ManagerBasedRLEnv
+    from isaaclab.sensors import FrameTransformer
 
 
 def object_reached_goal(
@@ -56,3 +58,61 @@ def object_reached_goal(
 
     # rewarded if the object is lifted above the threshold
     return distance < threshold
+
+
+def deformable_com_below_minimum(
+    env: ManagerBasedRLEnv,
+    minimum_height: float,
+    asset_cfg: SceneEntityCfg = SceneEntityCfg("deformable"),
+) -> torch.Tensor:
+    """Termination signal when the deformable's COM falls below ``minimum_height`` [m]."""
+    asset: DeformableObject = env.scene[asset_cfg.name]
+    com_z = wp.to_torch(asset.data.root_pos_w)[:, 2]
+    return com_z < minimum_height
+
+
+def deformable_outside_table_bounds(
+    env: ManagerBasedRLEnv,
+    x_bounds: tuple[float, float],
+    y_bounds: tuple[float, float],
+    asset_cfg: SceneEntityCfg = SceneEntityCfg("deformable"),
+) -> torch.Tensor:
+    """Terminate if any deformable nodal point leaves the table footprint.
+
+    Args:
+        env: The environment instance.
+        x_bounds: Allowed x-position range in the environment frame [m].
+        y_bounds: Allowed y-position range in the environment frame [m].
+        asset_cfg: The deformable object entity.
+
+    Returns:
+        Boolean tensor with shape ``(num_envs,)``.
+    """
+    asset: DeformableObject = env.scene[asset_cfg.name]
+    nodal_pos = wp.to_torch(asset.data.nodal_pos_w) - env.scene.env_origins.unsqueeze(1)
+    outside_x = (nodal_pos[..., 0] < x_bounds[0]) | (nodal_pos[..., 0] > x_bounds[1])
+    outside_y = (nodal_pos[..., 1] < y_bounds[0]) | (nodal_pos[..., 1] > y_bounds[1])
+    return torch.any(outside_x | outside_y, dim=1)
+
+
+def ee_below_minimum(
+    env: ManagerBasedRLEnv,
+    minimum_height: float,
+    ee_frame_cfg: SceneEntityCfg = SceneEntityCfg("ee_frame"),
+) -> torch.Tensor:
+    """Termination signal when the end-effector falls below ``minimum_height`` [m].
+
+    Height is measured in the environment frame (``z`` of the EE position with the env
+    origin subtracted), so the threshold is independent of the environment's xy offset.
+
+    Args:
+        env: The environment instance.
+        minimum_height: Minimum allowed EE height in the environment frame [m].
+        ee_frame_cfg: The end-effector frame entity.
+
+    Returns:
+        Boolean tensor with shape ``(num_envs,)``.
+    """
+    ee_frame: FrameTransformer = env.scene[ee_frame_cfg.name]
+    ee_z = wp.to_torch(ee_frame.data.target_pos_w)[..., 0, 2] - env.scene.env_origins[:, 2]
+    return ee_z < minimum_height
