--- source/isaaclab_tasks/isaaclab_tasks/manager_based/manipulation/lift_franka_soft/mdp/__init__.pyi
@@ -1,28 +1,0 @@
-# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
-# All rights reserved.
-#
-# SPDX-License-Identifier: BSD-3-Clause
-
-__all__ = [
-    "deformable_com_below_minimum",
-    "deformable_ee_distance",
-    "deformable_com_goal_distance",
-    "deformable_com_in_robot_root_frame",
-    "DeformableSampledPointsInRobotRootFrame",
-    "deformable_lifted",
-    "deformable_outside_table_bounds",
-    "ee_below_minimum",
-    "gripper_close_action",
-]
-
-from .observations import DeformableSampledPointsInRobotRootFrame, deformable_com_in_robot_root_frame
-from .rewards import (
-    deformable_com_below_minimum,
-    deformable_ee_distance,
-    deformable_com_goal_distance,
-    deformable_lifted,
-    deformable_outside_table_bounds,
-    ee_below_minimum,
-    gripper_close_action,
-)
-from isaaclab.envs.mdp import *
