+++ tools/environ_docs.py
@@ -35,7 +35,6 @@ _PHYSICS_BACKEND_MIRROR_NAMES = frozenset(
         "newton_kamino",
         "newton_mjwarp",
         "newton_mjwarp_vbd",
-        "newton_mjwarp_vdb",
         "ovphysx",
         "physx",
         *PresetTarget.all_legacy_aliases().keys(),
